need to add a field check in requests

+++ src/controllers/director.ts
@@ -2,25 +2,33 @@ import { MongoDriverError } from "mongodb";
 import { Director } from "../models/director";
 import { Movie } from "../models/movie";
 import { handlerError } from "./error";
+import { query, validationResult, matchedData, param } from "express-validator";
 
 const getDirectors = (req: any, res: any) => {
-  Director.find()
+  return Director.find()
     .then((item: any) => res.status(200).json(item))
     .catch((err) => handlerError(res, err));
 };
 
 const addOneDirector = (req: any, res: any) => {
-  const director = new Director(req.body);
-  director
-    .save()
-    .then((item: any) => res.status(200).json(item))
-    .catch((err) => handlerError(res, err));
+  const data = matchedData(req);
+  if (data.name !== "" && data.name !== undefined && data.name !== null) {
+    const director = new Director(data.name);
+    return director
+      .save()
+      .then((item: any) => res.status(200).json(item))
+      .catch((err) => handlerError(res, err));
+  }
+  res.send({ errors: `Field 'name' is empty` });
 };
 
 const deleteDirector = (req: any, res: any) => {
-  Director.findByIdAndDelete(req.params.id)
-    .then((item: any) => res.status(200).json(item))
-    .catch((err) => handlerError(res, err));
+  if (param("id").isMongoId()) {
+    return Director.findByIdAndDelete(req.params.id)
+      .then((item: any) => res.status(200).json(item))
+      .catch((err) => handlerError(res, err));
+  }
+  res.send({ errors: `Field 'id' is incorrect` });
 };
 
 const updateDirector = (req: any, res: any) => {
